+++ __tests__/services/users.service.spec.ts
@@ -1,116 +1,85 @@
 import { Container } from 'typedi';
 import { genSaltSync, hashSync } from 'bcrypt';
 import { UsersService } from '@services/users.service';
-import { API } from '../utils';
-import request from 'supertest';
 import { factory } from 'typeorm-seeding';
 import { User } from '@entities/user.entity';
-import app from '@app';
-import { HttpStatusCode } from '@constants/httpStatusCode';
 import { ErrorsMessages } from '@constants/errorMessages';
 import { getRepository, Repository } from 'typeorm';
-import { HashInvalidError } from '@exception/users/hashinvalid.error';
-import { HashExpiredError } from '@exception/users/hashexpired.error';
 import { mockUpdateResult } from '../utils/mocks';
 
 let usersService: UsersService;
 let user: User;
 let userRepository: Repository<User>;
 
-beforeAll(async () => {
-  usersService = Container.get(UsersService);
-});
-
-describe('compare password', () => {
-  let userPassword: string;
-
-  beforeEach(() => {
-    userPassword = hashSync('password', genSaltSync());
-  });
-
-  it('checks that the password matches', () => {
-    const hashedPassword = 'password';
-    const result = usersService.comparePassword({
-      password: hashedPassword,
-      userPassword
-    });
-    expect(result).toBeTruthy();
-  });
-
-  it("checks that the password don't match", () => {
-    const password = 'anotherpassword';
-    const result = usersService.comparePassword({ password, userPassword });
-    expect(result).toBeFalsy();
-  });
-});
-
-describe('verifying an account', () => {
-  beforeEach(async () => {
-    user = await factory(User)().create();
-  });
-
+describe('UsersService', () => {
   beforeAll(async () => {
     usersService = Container.get(UsersService);
-    userRepository = getRepository<User>(User);
   });
 
-  it('should throw error if the hash is invalid', async () => {
-    const verifyHash = '11111111-2222-3333-4444-555555555555';
-
-    jest.spyOn(userRepository, 'findOne')
-      .mockRejectedValueOnce(new Error(ErrorsMessages.HASH_NOT_VALID));
-
-    expect(usersService.verifyUser(verifyHash)).rejects.toThrowError(Error);
+  it('all dependencies should be defined', () => {
+    expect(usersService).toBeDefined();
   });
+  describe('comparePassword', () => {
+    let userPassword: string;
 
-  it('should throw error if the hash is expired', async () => {
-    const verifyHash = '11111111-2222-3333-4444-555555555555';
+    beforeEach(() => {
+      userPassword = hashSync('password', genSaltSync());
+    });
 
-    jest.spyOn(userRepository, 'findOne')
-      .mockRejectedValueOnce(new Error(ErrorsMessages.HASH_EXPIRED));
+    it('checks that the password matches', () => {
+      const hashedPassword = 'password';
+      const result = usersService.comparePassword({
+        password: hashedPassword,
+        userPassword
+      });
+      expect(result).toBeTruthy();
+    });
 
-    expect(usersService.verifyUser(verifyHash)).rejects.toThrowError(Error);
+    it("checks that the password don't match", () => {
+      const password = 'anotherpassword';
+      const result = usersService.comparePassword({ password, userPassword });
+      expect(result).toBeFalsy();
+    });
   });
+  describe('verifyUser', () => {
+    beforeEach(async () => {
+      user = await factory(User)().create();
+    });
 
-  it('should verify the user email', async () => {
-    user = await factory(User)().make();
-    jest.spyOn(userRepository, 'findOne')
-      .mockResolvedValueOnce(user);
-    jest.spyOn(userRepository, 'update')
-      .mockResolvedValueOnce(mockUpdateResult);
+    beforeAll(async () => {
+      usersService = Container.get(UsersService);
+      userRepository = getRepository<User>(User);
+    });
 
-    const userVerified = await usersService.verifyUser(user.verifyHash);
-    expect(userVerified).toHaveProperty('verified', true);
-    expect(userVerified).toHaveProperty('verifyHash', null);
-    expect(userVerified).toHaveProperty('hashExpiresAt', null);
-  });
+    it('should throw error if the hash is invalid', async () => {
+      const verifyHash = '11111111-2222-3333-4444-555555555555';
 
-  it('returns http code 500 with expired hash', async () => {
-    user = await factory(User)().create({ hashExpiresAt: new Date() });
-    const failingResponse = await request(app)
-      .get(`${API}/users/verify/?key=${user.verifyHash}`);
-    expect(failingResponse.status).toBe(HttpStatusCode.NOT_ACCEPTABLE);
-    expect(failingResponse.body).toStrictEqual(
-      expect.objectContaining(new HashExpiredError())
-    );
-  });
+      jest.spyOn(userRepository, 'findOne')
+        .mockRejectedValueOnce(new Error(ErrorsMessages.HASH_NOT_VALID));
 
-  it('returns http code 400 with invalid hash', async () => {
-    user.verifyHash = '11111111-2222-3333-4444-555555555555';
-    const failingResponse = await request(app)
-      .get(`${API}/users/verify/?key=${user.verifyHash}`);
-    expect(failingResponse.status).toBe(HttpStatusCode.BAD_REQUEST);
-    expect(failingResponse.body).toStrictEqual(
-      expect.objectContaining(new HashInvalidError())
-    );
-  });
+      expect(usersService.verifyUser(verifyHash)).rejects.toThrowError(Error);
+    });
 
-  it('returns http code 200 with valid params', async () => {
-    const response = await request(app)
-      .get(`${API}/users/verify/?key=${user.verifyHash}`);
-    expect(response.status).toBe(200);
-    expect(response.body.verified).toBe(true);
-    expect(response.body.hashExpiresAt).toBe(null);
-    expect(response.body.verifyHash).toBe(null);
+    it('should throw error if the hash is expired', async () => {
+      const verifyHash = '11111111-2222-3333-4444-555555555555';
+
+      jest.spyOn(userRepository, 'findOne')
+        .mockRejectedValueOnce(new Error(ErrorsMessages.HASH_EXPIRED));
+
+      expect(usersService.verifyUser(verifyHash)).rejects.toThrowError(Error);
+    });
+
+    it('should verify the user email', async () => {
+      user = await factory(User)().make();
+      jest.spyOn(userRepository, 'findOne')
+        .mockResolvedValueOnce(user);
+      jest.spyOn(userRepository, 'update')
+        .mockResolvedValueOnce(mockUpdateResult);
+
+      const userVerified = await usersService.verifyUser(user.verifyHash);
+      expect(userVerified).toHaveProperty('verified', true);
+      expect(userVerified).toHaveProperty('verifyHash', null);
+      expect(userVerified).toHaveProperty('hashExpiresAt', null);
+    });
   });
 });
